+++ HISTORY.rst
@@ -35,6 +35,7 @@ End-User Summary
 - Adding some UI documenation.
 - Fixed bug where case alignment stats were not properly imported.
 - Fixed bug where unfolding smallvariant details of a variant in a cohort that was not part of the base project caused a 404 error.
+- Fixed bug that prevented case import from API.
 
 Full Change List
 ================
@@ -77,6 +78,7 @@ Full Change List
 - Fixed bug where case alignment stats were not properly imported.
   Refactored case import in a sense that the new variant set gets activated when it is successfully imported.
 - Fixed bug where unfolding smallvariant details of a variant in a cohort that was not part of the base project caused a 404 error.
+- Fixed bug that prevented case import from API.
 
 -------
 v0.21.0